OpenAPI

+++ src/main/java/sorbie/jon/springwebopenapi/controller/Item.java
@@ -2,11 +2,11 @@ package sorbie.jon.springwebopenapi.controller;
 
 public class Item {
 
-    public Integer itemNumber;
+    public Long itemNumber;
     public String name;
     public String description;
 
-    public Item(Integer itemNumber, String name, String description) {
+    public Item(Long itemNumber, String name, String description) {
         this.itemNumber = itemNumber;
         this.name = name;
         this.description = description;
